requirements.txt
pillow~=9.4.0
transformers~=4.32.1